Apply styling to modal dialog and elements therein
Set margin auto on dialog in order to center it in viewport.
Applied max-width and max-height to accomodate for smaller devices.
Set darker background for modal dialog using ::backdrop to attract
more attention on the dialog content.

Applied Grid to form and placed elements in their desired positions.
Applied styling on form elements for better UI and UX.