Refactor radio element id and addBook(), attach event to submit button
The idea to split id name of radio input element was a dud. Therefore,
restored it to the way it was.

addBook() function now accesses radio element using 'name' attribute.

Attached event listener to 'Add Book' (submit) button of the form.
The button type has been changed to 'reset' because initially when the
form was submitted, the the input data was never lost. Therefore, to add
an auto-reset feature, the button type was altered.
Since we are not actually submitting the form to the server, there is
no need of the 'submit' type button.