Convert 'read' element to button and add toggle function
The read element has been changed to button as it describes the status
of completion of a book. Therefore, it needs to update status upon click.

toggleRead() function has been created which changes the read status
upon click.

The toggleRead() function is instantiated in addToDisplay() function
body. It attaches an event listener to each read button upon creation.

The code for setting up of read button in addToDisplay() function has
been altered to accomodate for setting up of background color to the
element.