Add function to remove book from library, justify script code
Added a remove button along with dynamic book elements added using
JS.

The removeBook() function is called within the addToDisplay() function
just like the toggleRead() function and is executed whenever a new book
is added to library.

The removeBook() function takes the remove button and the element
referencing the book (the parent of remove button) as its arguments
and attaches an event listener to the button which deletes the book
node upon click.

The script code has been justified, i.e., the code blocks of different
functions performing different tasks have been set in a particular order.
In short, the functions to open and close the dialog have been placed at
bottom in order to put addToDisplay(), toggleRead(), removeBook()
functions together since they are interconnected (as one is called inside
another.)