Refactor addToDisplay() and addBook() to change how 'read' is handled
Initially, the 'read' property was stored as a boolean value which made
the code messier and lengthy in addToDisplay() and served as an obstacle
in turning toggleRead() into an object prototype method.

Therefore, the values read now represent the values to be displayed on
screen, hence making things easier to handle.