Remove toggleRead() & recreate as prototype method, refactor execution code
The initial toggleRead() function did not actually change the value of the
read object variable.

Also, the standard way to define a function for an object is to define it
as method on the object property.
Since we also have to change the value of an object variable (read),
toggleRead() should definitely be an object prototype method rather than
a standalone function.

The book.prototype.toggleRead(readBtn) method takes in the DOM element
representing the 'read' object variable as an argument in order to change
its text content and background color.

The method is called via a click event listener which is attached in
addToDisplay() when the DOM element is created for the same.