Refactor .toggleRead() prototype method and DOM event listener
Initially, the prototype method .toggleRead() changed the object
variable (read) value as well as the text content & bg color of
DOM element.

However, this is not a good practice as "what happens in DOM stays
in DOM and what happens in object stays in object."
That is to say, if we were to call the method in console, it would
not be possible without providing it with a DOM element referencing
the read button.
This creates issues when changing values in an environment devoid
of webpage (like node).

Therefore, the method now only changes the value of variable in the
object.
The DOM element is updated inside the event listener called during
creation of DOM element in addToDisplay() function.