Refactor removeBook() and give unique ID to each object
Initially, the removeBook() function ONLY deleted the DOM element
of the book and didn't remove it from the myLibrary array storing all
the book objects. This was a serious issue as the book was deleted in
name only, but it was actually right there!

Therefore, in order to remove a specific object from myLibrary[], a
new object variable 'uniqueID' has been created which serves as a
unique identity for each created object.
The ID is assigned by creating a simple global variable declared and
assigned a value of 0; and each time an object is constructed, it is
passed as an argument for uniqueID of object and incremented by one.
This makes sure that each successive created object has a unique ID.
P.S. the global variable created is also named 'uniqueID', the same
as object variable.

The removeBook() function now takes in the uniqueID of current  book
object as an argument and loops through myLibrary[] to search for the
index containing the uniqueID; which when found, is replaced by the
object at the last index of the array and the last index is then
popped using .pop(); making sure that the object is completely
dereferenced.

The removeBook() no longer removes the DOM element, but only removes
the object from myLibrary[]. The DOM element is instead removed in
the event listener function. The reason is to keep the workings of
the myLibrary[] array and DOM elements separate in addition to the
reasons mentioned in the previous commit.