Add custom validation for empty form submission
Initially, the user was able to add books with no title, author name,
or pages even with the required attribute present.
This was because the submit button is actually of type 'reset' because
of which the auto-validation did not kick-in.

Therefore, we have added custom validation by returning a boolean
value from addBook() method of DisplayControl class. In absence
of title or author or pages info, the method returns false and the
error prompt is displayed to the user below the form input fields.

The error prompt has been coded in HTML and initially hidden using
CSS. It is made visible when an empty form is submitted.