Setup submit button type; add required attribute to input fields
The submit button was initially set to type="reset" in order to
reset the fields upon submit/cancel.
The submit button was not required as we were validating using pure
JS.

With this commit, however, we are utilising the Constraint Validation
API in order to validate and hence we require the submit button to be
of type="submit" in order to kick-off the automatic validation upon
submitting the form.
We have added the 'required' attribute in order to prevent the user
from submitting empty fields.